// src/services/models/GrowthModel.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type GrowthModel = {
    id?: string | null;
    infantId?: string;
    clinicalTeamId?: string;
    weight?: number;
    height?: number;
    headCircumference?: number;
    comment?: string | null;
    date?: string;
};